Split up kafka-influxdb main module into the KafkaInfluxDB class and helper methods
