--- kafka_influxdb/__init__.py
@@ -1,5 +1,0 @@
-__title__ = 'kafka_influxdb'
-from .version import __version__
-__author__ = 'Matthias Endler'
-__license__ = 'Apache License 2.0'
-__copyright__ = 'Copyright 2015, Matthias Endler under Apache License, v2.0'
