+++ kafka_influxdb/kafka_influxdb.py
@@ -1,17 +1,5 @@
-# Make it possible to call this file directly with python
-# See http://stackoverflow.com/a/19190695/270334
-if __name__ == '__main__' and __package__ is None:
-    from os import sys, path
-    sys.path.append(path.dirname(path.dirname(path.abspath(__file__))))
-
-import sys
 import logging
 import time
-from config import loader
-from encoder import load_encoder
-from reader import kafka_reader
-from writer import influxdb_writer
-from writer import kafka_sample_writer as benchmark
 
 
 class KafkaInfluxDB(object):
@@ -88,61 +76,3 @@ class KafkaInfluxDB(object):
 
     def get_config(self):
         return self.config
-
-
-def create_sample_messages(config):
-    print("Starting in benchmark mode. Stand by while creating sample messages.")
-    logging.info("Writing sample messages for benchmark to topic %s", config.kafka_topic)
-    bench = benchmark.KafkaSampleWriter(config)
-    bench.produce_messages()
-
-
-def main():
-    """
-    Setup consumer
-    """
-    config = loader.load_config()
-    logging.info("Connecting to Kafka broker at %s:%s", config.kafka_host, config.kafka_port)
-    if config.benchmark:
-        create_sample_messages(config)
-    start_consumer(config)
-
-
-def start_consumer(config):
-    """
-    Start metrics consumer
-    """
-    try:
-        reader = kafka_reader.KafkaReader(config.kafka_host,
-                                          config.kafka_port,
-                                          config.kafka_group,
-                                          config.kafka_topic)
-    except Exception as e:
-        logging.error("The connection to Kafka can not be established: %s. Please check your config.", e)
-        sys.exit(-1)
-
-    encoder = load_encoder(config.encoder)
-
-    logging.info("Connecting to InfluxDB at %s:%s", config.influxdb_host, config.influxdb_port)
-    try:
-        writer = influxdb_writer.InfluxDBWriter(config.influxdb_host,
-                                                config.influxdb_port,
-                                                config.influxdb_user,
-                                                config.influxdb_password,
-                                                config.influxdb_dbname,
-                                                config.influxdb_use_ssl,
-                                                config.influxdb_verify_ssl,
-                                                config.influxdb_timeout,
-                                                config.influxdb_use_udp,
-                                                config.influxdb_retention_policy,
-                                                config.influxdb_time_precision)
-    except Exception as e:
-        logging.error("The connection to InfluxDB can not be established: %s", e)
-        sys.exit(-2)
-
-    client = KafkaInfluxDB(reader, encoder, writer, config)
-    client.consume()
-
-
-if __name__ == '__main__':
-    main()
